requirements.txt
alabaster==0.7.16
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
boto3==1.34.55
botocore==1.34.55
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-currentuser==0.7.0
django-guardian==2.4.0
django-photologue==3.17
django-quill==0.1.0
django-quill-editor==0.1.40
django-simple-history==3.5.0
django-sortedm2m==3.1.1
django-storages==1.14.2
docutils==0.20.1
esbonio==0.16.4
ExifRead==3.0.0
httplib2==0.22.0
idna==3.6
imagesize==1.4.1
Jinja2==3.1.3
jmespath==1.0.1
lsprotocol==2023.0.1
lxml==5.1.0
MarkupSafe==2.1.5
mdToRst==1.1.0
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
pip-review==1.3.0
platformdirs==4.2.0
psycopg==3.1.18
pycodestyle==2.11.1
pycparser==2.21
pygls==1.3.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-openid==2.2.5
python-openid2==3.2
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
urllib3==2.0.7
vaticinator==0.0.7
watchdog==4.0.0
